尝试对下面最底部分进行背景去除，以及区域阈值分割,知道16位转到8位会丢失大部分精度

+++ realsense/d435_driver.h
@@ -49,11 +49,11 @@ class D435 {
   cv::Mat depth_data;
   double min_distance;
   std::vector<std::vector<cv::Point> > result;  //存放凸包
-  int thread1;
-  int thread2;
-  int thread3;
-  int thread4;
-  int thread5;
-  int thread6;
+  double thread1;
+  double thread2;
+  double thread3;
+  double thread4;
+  double thread5;
+  double thread6;
 };
 #endif  // REALSENSE_D435_DRIVER_H_
